requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-dotenv==1.0.1
openai>=1.14.0
yfinance>=0.2.50
requests==2.31.0
  